Add support for Skip ROM and different Write buffer length
If the Serial Number is not passed we use Skip ROM command to
communicate. (This work only if a single slave is present)

We also support a write buffer with size up to 8 bytes.
Now you can write a single byte if needed.

Now we can skip the READ verification process. If the PF bit fail we
force the verify process.
If the verify process fail we attempt to read again the scratchpad.